12/11 user-profile9

+++ src/controllers/videoController.js
@@ -18,6 +18,9 @@ export const getEdit=async(req,res)=>{
   if(video===null){
     return res.status(404).render("404",{pageTitle:"Video Not Found"});
   }
+  if(String(video.owner)!==String(req.session.user._id)){
+    return res.status(403).redirect("/");
+  }
   return res.render("edit",{pageTitle:`Edit "${video.title}"`,video})
 };
 export const postEdit=async(req,res)=>{
@@ -28,6 +31,9 @@ export const postEdit=async(req,res)=>{
   const video=await Video.exists({_id:id});
   if(!video){
     return res.status(404).render("404",{pageTitle:"Video Not Found"});
+  }
+  if(String(video.owner)!==String(req.session.user._id)){
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id,{
     title,description,hashtags:Video.formatHashtags(hashtags)
@@ -73,6 +79,13 @@ export const postUpload=async(req,res)=>{
 };
 export const deleteVideo=async(req,res)=>{
   const id=req.params.id;
+  const video=await Video.findById(id);
+  if(!video){
+    return res.status(404).render("404",{pageTitle:"Video Not Found"});
+  } 
+  if(String(video.owner)!==String(req.session.user._id)){
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 }
